test(api): 변경된 파일 구조에 맞춰 테스트 수정
