+++ api/api/v1/meal.test.ts
@@ -4,7 +4,7 @@ import express from 'express'
 import type { Request, Response } from 'express'
 import type { VercelRequest, VercelResponse } from '@vercel/node'
 import handler, { MealResponse } from './meal'
-import { validateDate, validateQueries, handleNeisStatus, formatMeal } from './meal'
+import { formatMeal } from './meal'
 import { startServer, resetServer, closeServer } from './__mocks__/mealServer'
 import {
   singleMealResponse,
@@ -136,199 +136,6 @@ describe('API', () => {
       expect(response.type).toBe('application/json')
       expect(response.body).toEqual(errorResponses.badQuery)
     })
-  })
-})
-
-describe('날짜 형식 확인', () => {
-  test('올바른 형식 (YYYY-MM-DD)', () => {
-    expect(validateDate('2024-03-20')).toBe(true)
-  })
-
-  test('올바른 형식 (YYYYMMDD)', () => {
-    expect(validateDate('20240320')).toBe(true)
-  })
-
-  test('잘못된 형식', () => {
-    expect(validateDate('2024-03-20T12:00:00')).toBe(false)
-    expect(validateDate('2024/03/20')).toBe(false)
-    expect(validateDate('2024.03.20')).toBe(false)
-  })
-
-  test('날짜가 아님', () => {
-    expect(validateDate('ㅁㄴㅇㄹ')).toBe(false)
-    expect(validateDate('')).toBe(false)
-  })
-})
-
-describe('쿼리 유효성 검사', () => {
-  test('모든 필드가 유효한 경우 (단일 날짜)', () => {
-    expect(
-      validateQueries('B10', '7010084', {
-        date: '2024-03-15',
-        startDate: undefined,
-        endDate: undefined,
-      }),
-    ).toBe(true)
-  })
-
-  test('모든 필드가 유효한 경우 (날짜 범위)', () => {
-    expect(
-      validateQueries('K10', '7872032', {
-        date: undefined,
-        startDate: '2024-03-18',
-        endDate: '2024-03-22',
-      }),
-    ).toBe(true)
-  })
-
-  test('필수 필드가 누락된 경우', () => {
-    expect(
-      validateQueries('', '7010084', {
-        date: '2024-03-15',
-        startDate: undefined,
-        endDate: undefined,
-      }),
-    ).toBe(false)
-
-    expect(
-      validateQueries('B10', '', {
-        date: '2024-03-15',
-        startDate: undefined,
-        endDate: undefined,
-      }),
-    ).toBe(false)
-  })
-
-  test('날짜 필드가 누락된 경우', () => {
-    expect(
-      validateQueries('B10', '7010084', {
-        date: undefined,
-        startDate: undefined,
-        endDate: undefined,
-      }),
-    ).toBe(false)
-  })
-
-  test('날짜 범위 중 하나만 있는 경우', () => {
-    expect(
-      validateQueries('B10', '7010084', {
-        date: undefined,
-        startDate: '2024-03-18',
-        endDate: undefined,
-      }),
-    ).toBe(false)
-
-    expect(
-      validateQueries('B10', '7010084', {
-        date: undefined,
-        startDate: undefined,
-        endDate: '2024-03-22',
-      }),
-    ).toBe(false)
-  })
-
-  test('단일 날짜와 날짜 범위가 동시에 있는 경우', () => {
-    expect(
-      validateQueries('B10', '7010084', {
-        date: '2024-03-15',
-        startDate: '2024-03-18',
-        endDate: '2024-03-22',
-      }),
-    ).toBe(false)
-  })
-
-  test('시작일이 종료일보다 늦은 경우', () => {
-    expect(
-      validateQueries('B10', '7010084', {
-        date: undefined,
-        startDate: '2024-03-22',
-        endDate: '2024-03-18',
-      }),
-    ).toBe(false)
-  })
-
-  test('잘못된 형식의 필드', () => {
-    expect(
-      validateQueries('B1', '7010084', {
-        date: '2024-03-15',
-        startDate: undefined,
-        endDate: undefined,
-      }),
-    ).toBe(false)
-
-    expect(
-      validateQueries('B10', 'ABC', {
-        date: '2024-03-15',
-        startDate: undefined,
-        endDate: undefined,
-      }),
-    ).toBe(false)
-
-    expect(
-      validateQueries('B10', '7010084', {
-        date: '2024/03/15',
-        startDate: undefined,
-        endDate: undefined,
-      }),
-    ).toBe(false)
-  })
-})
-
-describe('나이스 상태 코드 처리', () => {
-  test('ERROR-300 처리', () => {
-    const result = handleNeisStatus('ERROR-300')
-    expect(result).toEqual({
-      status: 400,
-      body: {
-        error: {
-          code: 400,
-          message: '쿼리가 누락되었습니다.',
-        },
-      },
-    })
-  })
-
-  test('ERROR-337 처리', () => {
-    const result = handleNeisStatus('ERROR-337')
-    expect(result).toEqual({
-      status: 429,
-      body: {
-        error: {
-          code: 429,
-          message: '오늘 호출 횟수를 초과했습니다.',
-        },
-      },
-    })
-  })
-
-  test('기타 오류 처리', () => {
-    const errorCodes = [
-      'ERROR-290',
-      'ERROR-310',
-      'ERROR-333',
-      'ERROR-336',
-      'ERROR-500',
-      'ERROR-600',
-      'ERROR-601',
-      'INFO-300',
-    ]
-
-    for (const code of errorCodes) {
-      expect(handleNeisStatus(code)).toEqual({
-        status: 500,
-        body: {
-          error: {
-            code: 500,
-            message: '데이터를 불러오는 데 실패했습니다.',
-          },
-        },
-      })
-    }
-  })
-
-  test('정상 상태 코드 처리', () => {
-    const result = handleNeisStatus('INFO-000')
-    expect(result).toBeNull()
   })
 })
